change nonce to int

+++ providers/liquidity.go
@@ -58,7 +58,7 @@ func (lp *LocalProvider) GetQuote(q types.Quote, gas uint64, gasPrice big.Int) *
 	q.TimeForDeposit = 3600
 	q.CallTime = 3600
 	q.PenaltyFee = *big.NewInt(10)
-	q.Nonce = uint(rand.Uint64())
+	q.Nonce = rand.Int()
 	cost := big.NewInt(0).Mul(&gasPrice, big.NewInt(int64(gas)))
 	fee := cost.Div(cost, big.NewInt(33))
 	finalCost := cost.Add(cost, fee)
